+++ paper.md
@@ -16,7 +16,7 @@ authors:
   affiliation: 1
 
 affiliations:
-- name: Department of Music, New York University
+- name: Department of Music, New York University, New York, New York 10003
   index: 1
   
 date: 23 September 2021
