+++ paper.md
@@ -29,7 +29,7 @@ bibliography: paper.bib
 EDO.js is a JavaScript library for generating, analyzing, and visualizing pitch-collections within any tuning system of Equal 
 Divisions of an Octave (EDO). The current focus is on psychological experiments pertaining to musical 
 scale perception and pitch-centric music research more broadly. EDO.js implements major algorithms presented in the
-music theory and music cognition literature [e.g., @Balzano:1982; @clampitt; @rahn1991coordination; @vassilakis2001auditory; lerdahl2001tonal; @carey2007coherence], alongside a wide array of additional tools for analyzing and manipulating midi-data and melodies in the framework of musical set theory 
+music theory and music cognition literature [e.g., @Balzano:1982; @clampitt; @rahn1991coordination; @vassilakis2001auditory; @lerdahl2001tonal; @carey2007coherence], alongside a wide array of additional tools for analyzing and manipulating midi-data and melodies in the framework of musical set theory 
 [@forte1973structure] and diatonic set theory.
 
 Two classes are at the core of the package. The EDO class, and a daughter class, the Scale class. The EDO class and the Scale class conceptually diverge only on a single aspect: pitches in the Scale class are regarded as pitch-classes (exist only within an octave) and they have an enharmonic meaning. However, in the EDO class pitches are regarded as specific, and are not confined to a single octave. That is to say, while the Scale class treats C3 and C4 simply as “C”, the EDO class would regard them as separate entities. As such, the Scale class is better suited to engage with scales, their properties, and their structures, while the EDO class is an implementation of useful functions for engaging with more abstract musical structures.
