### 2021-09-28 #3760 @bmcfee
:wave: per offline conversation with @kthyng - I can edit this one.  I technically have a [COI](https://joss.theoj.org/about#ethics) due to being at the same institution as the author (NYU), but as we're in completely separate units (GSAS and Steinhardt, respectively), this shouldn't be a problem.
