### 2021-09-28 #3760 @danielskatz
On behalf of JOSS, we agree that this COI can be waived due to @bmcfee really being the only editor with the right background for this, and the fact that the units are separate.
