### 2021-09-30 #3784 @whedon
[REVIEW]: EDO.js: A comprehensive JavaScript library for interaction with musical set theory in any tuning
**Submitting author:** @MichaelSel (<a href="http://orcid.org/0000-0002-6061-0674">Michael Seltenreich</a>)
**Repository:** <a href="https://github.com/MichaelSel/edoJS" target ="_blank">https://github.com/MichaelSel/edoJS</a>
**Version:** v1.2.13

- [x] #3826
