### 2021-10-12 MichaelSel/edoJS#1 @napulen
[Paper] Missing section comparing against other software packages
This issue is related to the following review: https://github.com/openjournals/joss-reviews/issues/3784#event-5391142283

According to the guidelines of JOSS papers, a *description of how this software compares to other commonly-used packages in this research area* is missing.

I know of at least the existence of [music21](https://web.mit.edu/music21/), which overlaps on several aspects with `edo.js`, and has been [ported to JavaScript](https://web.mit.edu/music21/music21j/doc/index.html) in the past.

Another interesting thing for this project would be to include a line or two about how it could potentially complement tools like [vexflow](https://www.vexflow.com/), [Verovio](https://www.verovio.org/index.xhtml), or the [Verovio Humdrum Viewer](https://verovio.humdrum.org/).

As a reviewer, I'd accept the claim that all of these tools are tangential and `edo.js` is the first tool of its kind (e.g., to the best of my knowledge, no other analysis library has been written in pure JavaScript), but I'd still want the section in the paper stating that.
