### 2021-10-12 MichaelSel/edoJS#2 @napulen
[Paper] Is this portion of the paper considered "documentation"? If so, remove
This issue is related to the following review: https://github.com/openjournals/joss-reviews/issues/3784

Following the JOSS review criteria, the documentation mentions that

> the paper should not include software documentation such as API (Application Programming Interface) functionality, as this should be outlined in the software documentation.

In the paper, there are two subsections that could potentially be considered software documentation:
- The EDO class
- The Scale class

I am generally okay with having these subsections in the paper, but asking @albincorreya and @bmcfee for their input.
