### 2021-10-12 MichaelSel/edoJS#3 @napulen
[README] Missing "statement of need" in README
This issue is related to the following review: https://github.com/openjournals/joss-reviews/issues/3784

As per the JOSS guidelines, the README should have a ["statement of need"](https://joss.readthedocs.io/en/latest/review_criteria.html#a-statement-of-need).

I assume the "statement of need" to be the current introductory paragraph in the README, which states:
*A JavaScript library for pitch set manipulation in any EDO tuning system.*

I'd add the following information to this paragraph:

- Unpack the acronym (i.e., Equal Divisions of an Octave (EDO))
- Make the introduction a bit friendlier to a non-specialist audience who is just seeing the repo for the first time. For example
  - Briefly explain why pitch set manipulation is useful
  - What kind of problems can be approached by using this library
  - Who is the target audience of the software

The summary in the paper already addresses most of these points. A simple solution would be to repeat the summary of the paper in the README.

At the very minimum, please unpack the acronym. Even as someone very familiar with Western music theory, I had to first search `EDO` to know what it meant.
