### 2021-10-12 41b7cc4
Updated README to elaborate statement of need
