### 2021-10-12 MichaelSel/edoJS#5 @napulen
[Code] Missing package.json or dependency control
This issue is related to the following review: https://github.com/openjournals/joss-reviews/issues/3784

From the main script, it seems that this `edo.js` requires three third-party dependencies:

- `fs`
- `xml2js`
- `midi-parser-js`

According to the [JOSS guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions), these should be indicated in a `package.json` or equivalent resource
