### 2021-10-12 MichaelSel/edoJS#6 @napulen
[README] Missing community guidelines for contribution, issues, and support
This issue is related to the following review: https://github.com/openjournals/joss-reviews/issues/3784

According to the [JOSS guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines), the README should have some community guidelines for the following things:

- [x] Contributing to the library
- [x] Reporting an issue
- [x] Obtaining support

Considering that this is a relatively small single-contributor library, I'd be okay assuming that the third point is not necessary. However, the guidelines for contribution and reporting issues would be very useful.
