### 2021-10-12 MichaelSel/edoJS#5 @MichaelSel
Added dependencies
