### 2021-10-12 MichaelSel/edoJS#7 @napulen
[Documentation] General feedback on the API documentation
This issue is related to the following review: https://github.com/openjournals/joss-reviews/issues/3784

I was playing around with a few functions of the library on my browser (e.g., [chord_quality](https://michaelsel.github.io/edoJS/Scale_count.html#.chord_quality)).

First of all, kudos for providing at least one example (with output) of every function! This was very useful when playing with the library.

Secondly, here are a few comments for improving the documentation:

- Please unpack the acronyms: It is clear for me to read something like `PCs`, but relying on unexplained acronyms makes the text harder to approach. I'd prefer to have each acronym expanded at least the first time it is used in a given page. This would make the documentation easier to approach.
- In the paper, there is a claim about `edo.js` implementing algorithms from the music research literature[^1], however, in the documentation, the relationship between major algorithms in music research and custom functions is not clearly presented. In most music-processing libraries, a reference to the source publication/algorithm is shown in the API documentation. See, for example, the `References` sections of [librosa](https://librosa.org/doc/latest/generated/librosa.decompose.hpss.html), [madmom](https://madmom.readthedocs.io/en/v0.16.1/modules/audio/chroma.html), or [essentia](https://essentia.upf.edu/reference/std_MFCC.html). Something similar here would be very useful.
- The wording and length of the explanations is sometimes insufficient
  - For example, it took me a bit to figure out what exactly `chord_quality` does
  - I think, for this particular one, "specified in ~PCs~ semitones above the root" is a better description, because the PCs will be transposed for every root it is comparing against, but the number of semitones above the root is constant (i.e., as provided by the user), right? This is why it took me a while to figure it out. The wording here makes a substantial difference, in my opinion.
  - In a similar line, what is the difference between `chord_quality([4])` and `M3s()`? Is it just a shortcut? This would be useful in the documentation

Take all these, especially the third point, as suggestions. I understand it would require a significant amount of work to implement.

[^1]: The precise claim is: *EDO.js implements major algorithms presented in the music theory and music
cognition literature (e.g., Balzano, 1982; Carey, 2007; Carey & Clampitt, 1989; Lerdahl &10
others, 2001; Rahn, 1991; Vassilakis, 2001)*
