### 2021-10-12 MichaelSel/edoJS#8 @napulen
[Code] Automated tests?
This issue is related to the following review: https://github.com/openjournals/joss-reviews/issues/3784

According to the [JOSS guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#tests), the software is expected to have automated tests.

I think, given that every function in the API seems to have an example with expected output, that this is not very far from implemented. However, it'd be great if an automated test could run the actual/expected outputs of each function.
