### 2021-10-12 MichaelSel/edoJS#9 @napulen
[References] A few minor corrections
This issue is related to the following review: https://github.com/openjournals/joss-reviews/issues/3784

On the `paper.bib` file:

- On Lerdahl (2001), are there really other authors involved? I have a copy of *Tonal Pitch Space* and based on what I see, it is solely written by Lerdahl.
- For consistency, make the indentation (i.e., first entry) and format of the identifiers (`<author><year><firstword>`) equal in all bibtex entries.
