### 2021-10-12 0f26fe7
Updated dependecies and removed duplications
