+++ edo.js
@@ -479,7 +479,7 @@ class EDO {
          * @example
          * let edo = new EDO(17) // define a tuning with 17 divisions of the octave
          * edo.convert.interval_to_cents(6)
-         * //returns 423.5294117647059*/
+         * //returns 600*/
         interval_to_cents: (interval) => {
             return this.cents_per_step * interval
         },
@@ -541,9 +541,10 @@ class EDO {
             let pcs = [0]
 
             intervals.forEach((interval) => {
-                pcs.push(interval + pcs[pcs.length - 1])
+                pcs.push((interval + pcs[pcs.length - 1]))
             })
-            return pcs
+
+            return this.scale(pcs).pitches
         },
 
         /** Given a list of midi notes, returns a list of intervals
@@ -639,7 +640,7 @@ class EDO {
          * @memberOf EDO#convert
          * @example
          * let edo = new EDO(12) // define a tuning system with 12 divisions of the octave
-         * edo.convert.midi_to_name(4)
+         * edo.convert.pc_to_name(4)
          * //returns "E"
          * */
         pc_to_name: (pc) => {
@@ -660,7 +661,7 @@ class EDO {
             }
             if (this.edo != 12) return pc
             if(Array.isArray(pc)) return pc.map(p=>this.convert.pc_to_name(p))
-            return PC[pc]
+            return PC[pc].trim()
         },
 
         /** Normalizes any input to include PC only (to ignore octave displacement)
@@ -985,10 +986,11 @@ class EDO {
          * edo.get.coordinates([0,3,7]) //minor triad
          * //returns
          * [
-         *  [ 0, 0.56418958354776 ],
-         *  [ 0.5043836053298857, -0.25279846688265545 ],
-         *  [ 0.26388191608137657, -0.4986744635020088 ]
+         *  [0,0.56418958354776],
+         *  [0.56418958354776,3.454664838020213e-17],
+         *  [-0.2820947917738801,-0.48860251190292314]
          * ]
+         *
          */
         coordinates: (pitch,circle_center = [0,0],r=0.56418958354776	) => {
             if(Array.isArray(pitch)) return pitch.map(p=>this.get.coordinates(p,circle_center,r))
@@ -1096,11 +1098,7 @@ class EDO {
          * @example
          * edo.get.combinations([1,3,5,7],2)
          * //returns
-         * [
-         *   { motive: [ -1, 1 ], incidence: 2 }, //going a half-step down, then up appears twice
-         *   { motive: [ -1 ], incidence: 2 }, //going a half-step down appears twice
-         *   { motive: [ 1 ], incidence: 2 } //going a half-step up appears twice
-         * ]
+         * [[1,3],[3,1],[1,5],[5,1],[1,7],[7,1],[3,5],[5,3],[3,7],[7,3],[5,7],[7,5]]
          */
         combinations: (set, k) => {
             let combs = this.get.n_choose_k(set,k)
@@ -1554,7 +1552,7 @@ class EDO {
          * @returns {Number}
          * @memberOf Scale#get
          * @example
[…]
