### 2021-10-13 MichaelSel/edoJS#8 @MichaelSel
This is something that's been on my mind. It's quite tricky to test. I can create deterministic tests where I check for the expected output for a given input for every function.
However, this is no guarantee that the code will work correctly with other inputs. Of course it would have been ideal to generate automated testing with random inputs, but there's no way to cross-check that the outputs are actually correct.

Another challenge is that some of the functions do not return deterministic outputs (random generators and such), and therefore, there's no way to verify their output.

I feel quite out of my comfort zone as I do not have vast experience in testing, and I would very much appreciate some guidance if you're willing to offer it.

Thanks!
