### 2021-10-13 ebd0018
Updated dependecies and removed duplications
