### 2021-10-13 192821f
Updated dependecies and removed duplications
