+++ edo.js
@@ -7209,100 +7209,99 @@ try {
 }
 
 
-function test_all() {
-     const JS = JSON.stringify
-     let edo = new EDO(12) // define a tuning system with 12 divisions of the octave
+// function test_all() {
+//      const JS = JSON.stringify
+//      let edo = new EDO(12) // define a tuning system with 12 divisions of the octave
+//
+//     const run_it = function (func_name,func,func_params, expected_return,compact=true) {
+//          const result = JS(func(...func_params))
+//         expected_return = JS(expected_return)
+//         const pass_fail = (result==expected_return)?"pass":"fail"
+//         if(pass_fail=="fail") throw new Error(func_name + " failed to pass automated test.\n Expected: " + expected_return + ", but got " + result)
+//         if(compact) console.log("Function: " + func_name + " pass/fail:", pass_fail )
+//         else console.log("Function: " + func_name +"\nparams:",func_params,"\nexpected:", expected_return,"\ngot:", result,"\npass/fail:", pass_fail )
+//     }
+//
+//
+//     let edo_convert = {
+//         cents_to_ratio:[edo.convert.cents_to_ratio,[700],1.4983070768766815],
+//         cents_to_simple_ratio:[edo.convert.cents_to_simple_ratio,[700],{"cents":1901.9550008653873,"cents_in_octave":701.9550008653873,"value":3,"diff_in_octave":-1.9550008653873192,"ratio":"3:1","original": 700}],
+//         interval_to_cents:[edo.convert.interval_to_cents,[6],600],
+//         interval_to_ratio:[edo.convert.interval_to_ratio,[7],1.4983070768766815],
+//         intervals_to_pitches:[edo.convert.intervals_to_pitches,[[2,3]],[ 0, 2, 5 ]],
+//         intervals_to_scale: [edo.convert.intervals_to_scale,[[2, 2, 1, 2, 2, 2, 1]],[0,2,4,5,7,9,11]],
+//         midi_to_intervals: [edo.convert.midi_to_intervals, [[60,64,57,61]], [ 4, -7, 4 ]],
+//         midi_to_name: [edo.convert.midi_to_name, [[60,62]],["C4","D4"]],
+//         midi_to_freq: [edo.convert.midi_to_freq,[[69,70]],[ 440, 466.1637615180899 ]],
+//         // name_to_scale: [edo.convert.name_to_scale,['12-1387']]
+//         pc_to_name:[edo.convert.pc_to_name,[4],"E"],
+//         pitches_to_PCs: [edo.convert.pitches_to_PCs,[[0,2,12,-2,7]],[0,2,0,10,7]],
+//         ratio_to_cents: [edo.convert.ratio_to_cents,[5/4],386.3137138648348],
+//         ratio_to_interval: [edo.convert.ratio_to_interval,[5/4,20],[4]],
+//         to_steps: [edo.convert.to_steps,[[0,2,4,5,7,9,11]],[ 2, 2, 1, 2, 2, 2 ]]
+//
+//     }
+//     let edo_count = {
+//         common_tones: [edo.count.common_tones, [[1,2,4],[2,3,4,5]], 2],
[…]
