### 2021-10-13 MichaelSel/edoJS#7 @MichaelSel
Thanks for the thoughtful comments!

- I will unpack all of the acronyms and let you know when it's done.

- As for citing: a lot of the functions implemented in the library are either novel, or "organic", such that they do not have documented authorship. The ones that do, often carry the author's name in the name of the function, and those have the reference in the description.
For instance in `scale.get.lerdahl_attraction()` , `edo.get.maximal_carey_coherence_failures()`, `scale.count.rahn_differences()`, etc. all include a citation to the paper that describes the algorithm.

- Lastly, I can see where the documentation may be confusing and I will make effort to make it clearer. Specifically, your comments make me think that the **tuning** aspect of the library is not emphasized enough.

Let me clarify here first, and hopefully, I'll find a way to make it clearer in the documentation as well.
I see why you prefer  "specified in semitones" over "specified in PCs". However, those phrases are interchangeable **only** in 12EDO.

In 12EDO, the pitch-classes [4 9] (like you say) represent 4 and 9 semitones above 0.
**However**, in 24EDO the pitch-classes [4 9] represent 4 and 9 **quarter-tones** above 0 (as the octave in this tuning is divided into 24 equal parts, not 12, hence quarter-tones rather than semitones).
A more extreme example: in 1200EDO, where the octave will be cut into 1200 parts, [4 9] will represent 4 and 9 **cents** above 0. For that reason, I must keep the less intuitive, but more accurate "specified in PCs" wording. Where the pitch classes have different sizes in different tunings.

 Similarly, the difference between chord_quality([4]) and M3s() is a practical and a conceptual one.  chord_quality([4]) "thinks" in terms of set-theory and M3s are calculated by finding the closest pitch-classes to an **acoustic** ratio of 5:4 (with a default tolerance of 20 cents)

So if we had 3 tuning contexts:

```
let edo12 = new EDO(12) //12 equal divisions of the octave - Standard
let edo17 = new EDO(17) //17 equal divisions of the octave
let edo31 = new EDO(31) //31 equal divisions of the octave - Turkish tuning

//Let's define a basic (seemingly) "major" scale in all contexts
let scale12 = edo12.scale([0,2,4,5,7,9,11])
let scale17 = edo17.scale([0,2,4,5,7,9,11])
let scale31 = edo31.scale([0,2,4,5,7,9,11])

//in the scale [0, 2, 4, 5, 7, 9 ,11]   [4] can be found in [0 4]    [5 9]    [7 11]
//Because the pitch classes are the same, the following output is not surprising

scale12.count.chord_quality([4]) // returns 3
scale17.count.chord_quality([4]) // returns 3
scale31.count.chord_quality([4]) // returns 3
```

But, in edo12 a Major 3rd corresponds to an interval of size [4], in edo17 there's no interval within 20 cents of a 5:4 ratio, and in edo31 a Major 3rd corresponds with an interval of size [10].

Therefore:
```
scale12.count.M3s() // as expected returns 3
scale17.count.M3s() // returns 0 because acoustically there's no major 3rds in 17EDO
scale 31.count.M3s() // returns 0 as well, because while 31EDO has Major 3rds, the given pitches in THIS tuning do not contain a major 3rd ([10])

```

As you can see, this is not so trivial. I will try to make this clearer in the documentation.
