+++ CONTRIBUTING.md
@@ -1,0 +1,5 @@
+`EDO.js` would love some community contributions. The [project](https://github.com/MichaelSel/edoJS) is hosted at GitHub.
+
+## Submitting Pull Requests ##
+
+Anyone is welcome to contribute. Please open an issue to propose a feature or report a bug before raising a pull request.
