+++ CONTRIBUTING.md
@@ -1,5 +1,12 @@
 `EDO.js` would love some community contributions. The [project](https://github.com/MichaelSel/edoJS) is hosted at GitHub.
 
+
+## Issue Tickets ##
+
+Currently, there are no special instructions for creating an issue. For bugs, feature suggestions, and general discussions, feel free to create an issue. 
+
+
 ## Submitting Pull Requests ##
 
 Anyone is welcome to contribute. Please open an issue to propose a feature or report a bug before raising a pull request.
+At this time, there is not a well-established style-guide. However, make sure any contribution is well documented, and hopefully includes at least one example along with the expected return. 
