+++ paper.bib
@@ -49,7 +49,7 @@
 
 @book{lerdahl2001tonal,
   title={Tonal pitch space},
-  author={Lerdahl, Fred and others},
+  author={Lerdahl, Fred},
   year={2001},
   publisher={Oxford University Press, USA}
 }
