+++ paper.bib
@@ -1,22 +1,22 @@
 
 
 @book{Balzano:1982,
-  	url = {https://doi.org/10.1007/978-1-4684-8917-0_17},
   	Author = {Balzano, J. G.},
+  	Year = 1982
+  	Title = {{The Pitch Set as a Level of Description for Studying Musical Pitch Perception}},
+  	url = {https://doi.org/10.1007/978-1-4684-8917-0_17},
   	Booktitle = {Balzano G.J. (1982) The Pitch Set as a Level of Description for Studying Musical Pitch Perception. In: Clynes M. (eds) Music, Mind, and Brain. Springer, Boston, MA. https://doi.org/10.1007/978-1-4684-8917-0_17},
   	Publisher = {Springer, Boston, MA},
-  	Title = {{The Pitch Set as a Level of Description for Studying Musical Pitch Perception}},
-  	Year = 1982
 }
 
 @article{clampitt,
     author = {Carey, Norman and Clampitt, David},
+    year = {1989},
     title = "{Aspects of Well-Formed Scales}",
     journal = {Music Theory Spectrum},
     volume = {11},
     number = {2},
     pages = {187-206},
-    year = {1989},
     month = {10},
     abstract = "{Pentatonic, diatonic, and chromatic scales share the same underlying structure, that of the well-formed scale. Well-formedness is defined in terms of a relationship between the order in which a single interval generates the elements of a pitch-class set and the order in which those elements appear in a scale. Another characterization provides a recursive procedure for organizing all well-formed scales into hierarchies. Finally, well-formed scales are defined in terms of scale-step measure, and aspects of the diatonic set are examined.}",
     issn = {0195-6167},
@@ -26,49 +26,49 @@
 }
 
 @article{rahn1991coordination,
-  title={Coordination of interval sizes in seven-tone collections},
   author={Rahn, Jay},
+  year={1991},
+  title={Coordination of interval sizes in seven-tone collections},
   journal={Journal of Music Theory},
   volume={35},
   number={1/2},
   pages={33--60},
-  year={1991},
   publisher={JSTOR}
 }
 
 @article{vassilakis2001auditory,
-  title={Auditory roughness estimation of complex spectra—Roughness degrees and dissonance ratings of harmonic intervals revisited},
   author={Vassilakis, Pantelis},
+  year={2001},
+  title={Auditory roughness estimation of complex spectra—Roughness degrees and dissonance ratings of harmonic intervals revisited},
   journal={The Journal of the Acoustical Society of America},
   volume={110},
   number={5},
   pages={2755--2755},
-  year={2001},
   publisher={Acoustical Society of America}
 }
 
 @book{lerdahl2001tonal,
-  title={Tonal pitch space},
   author={Lerdahl, Fred},
   year={2001},
+  title={Tonal pitch space},
   publisher={Oxford University Press, USA}
 }
 
 @article{carey2007coherence,
-  title={Coherence and sameness in well-formed and pairwise well-formed scales},
   author={Carey, Norman},
+  year={2007},
+  title={Coherence and sameness in well-formed and pairwise well-formed scales},
   journal={Journal of Mathematics and Music},
   volume={1},
   number={2},
   pages={79--98},
-  year={2007},
   publisher={Taylor \& Francis}
 }
 
 @book{forte1973structure,
-  title={The structure of atonal music},
   author={Forte, Allen},
-  volume={304},
   year={1973},
+  title={The structure of atonal music},
+  volume={304},
   publisher={Yale University Press}
 }
