+++ paper.bib
@@ -1,28 +1,26 @@
-
-
 @book{Balzano:1982,
-  	Author = {Balzano, J. G.},
-  	Year = 1982
-  	Title = {{The Pitch Set as a Level of Description for Studying Musical Pitch Perception}},
-  	url = {https://doi.org/10.1007/978-1-4684-8917-0_17},
-  	Booktitle = {Balzano G.J. (1982) The Pitch Set as a Level of Description for Studying Musical Pitch Perception. In: Clynes M. (eds) Music, Mind, and Brain. Springer, Boston, MA. https://doi.org/10.1007/978-1-4684-8917-0_17},
-  	Publisher = {Springer, Boston, MA},
+  Author = {Balzano, J. G.},
+  Year = 1982
+  Title = {{The Pitch Set as a Level of Description for Studying Musical Pitch Perception}},
+  url = {https://doi.org/10.1007/978-1-4684-8917-0_17},
+  Booktitle = {Balzano G.J. (1982) The Pitch Set as a Level of Description for Studying Musical Pitch Perception. In: Clynes M. (eds) Music, Mind, and Brain. Springer, Boston, MA. https://doi.org/10.1007/978-1-4684-8917-0_17},
+  Publisher = {Springer, Boston, MA},
 }
 
 @article{clampitt,
-    author = {Carey, Norman and Clampitt, David},
-    year = {1989},
-    title = "{Aspects of Well-Formed Scales}",
-    journal = {Music Theory Spectrum},
-    volume = {11},
-    number = {2},
-    pages = {187-206},
-    month = {10},
-    abstract = "{Pentatonic, diatonic, and chromatic scales share the same underlying structure, that of the well-formed scale. Well-formedness is defined in terms of a relationship between the order in which a single interval generates the elements of a pitch-class set and the order in which those elements appear in a scale. Another characterization provides a recursive procedure for organizing all well-formed scales into hierarchies. Finally, well-formed scales are defined in terms of scale-step measure, and aspects of the diatonic set are examined.}",
-    issn = {0195-6167},
-    doi = {10.2307/745935},
-    url = {https://doi.org/10.2307/745935},
-    eprint = {https://academic.oup.com/mts/article-pdf/11/2/187/3800626/11-2-187.pdf},
+  author = {Carey, Norman and Clampitt, David},
+  year = {1989},
+  title = "{Aspects of Well-Formed Scales}",
+  journal = {Music Theory Spectrum},
+  volume = {11},
+  number = {2},
+  pages = {187-206},
+  month = {10},
+  abstract = "{Pentatonic, diatonic, and chromatic scales share the same underlying structure, that of the well-formed scale. Well-formedness is defined in terms of a relationship between the order in which a single interval generates the elements of a pitch-class set and the order in which those elements appear in a scale. Another characterization provides a recursive procedure for organizing all well-formed scales into hierarchies. Finally, well-formed scales are defined in terms of scale-step measure, and aspects of the diatonic set are examined.}",
+  issn = {0195-6167},
+  doi = {10.2307/745935},
+  url = {https://doi.org/10.2307/745935},
+  eprint = {https://academic.oup.com/mts/article-pdf/11/2/187/3800626/11-2-187.pdf},
 }
 
 @article{rahn1991coordination,
