### 2021-10-13 ce70dc3
Added some Unit tests
