+++ edo.js
@@ -2309,7 +2309,7 @@ class EDO {
          * @example
          * let edo = new EDO(12) //Create a tuning context
          * edo.get.normal_order([0,2,4,5,7,9,11])
-         * //returns [0, 1, 2, 3, 5, 6, 8, 10]*/
+         * //returns [ 0, 1, 3, 5, 6, 8, 10 ]*/
         normal_order: (lst, cache = false) => {
             if (lst.length == 0) return []
             let edo = this.edo
@@ -2592,8 +2592,9 @@ class EDO {
          * edo.get.pitch_distribution([0,12,0,12,7,0])
          * //returns
          * [
-         *  { note: 0, rate: 0.8333333333333334 },
-         *  { note: 7, rate: 0.16666666666666666 }
+         *  {"note": 0,"rate": 0.5},
+         *  {"note": 12,"rate": 0.3333333333333333},
+         *  {"note": 7,"rate": 0.16666666666666666}
          * ]
          *
          */
@@ -2858,11 +2859,14 @@ class EDO {
          * @example
          * let edo = new EDO(12) // create a tuning context
          * edo.get.ratio_approximation(7)
-         * //returns { ratio: '3/2', cents_offset: -1.955000865387433, decimal: 1.5 }
+         * //returns {
+         *  "cents_offset": -1.955000865387433
+         *  "decimal": 1.5
+         *  "log_position": 0.5849625007211562
+         *  "octave": 1
+         *  "ratio": "3:2"
+         *  }
          *
-         * let edo = new EDO(17) // Notice this is 17 divisions
-         * console.log(edo.get.ratio_approximation(3))
-         * //returns {ratio: '17/15', cents_offset: -4.921988887832043, decimal: 1.1333333333333333}
          *
          */
         ratio_approximation: (interval, limit = 17) => {
@@ -2913,8 +2917,8 @@ class EDO {
          * @memberOf EDO#get
          * @example
          * let edo = new EDO(12) // define a tuning system
-         * edo.get.rotations([0,2,4,5,7],2) //returns [4,5,7,0,2]
-         * edo.get.rotations([0,2,4,5,7],-1) //returns [7,0,2,4,5]
+         * edo.get.rotated([0,2,4,5,7],2) //returns [4,5,7,0,2]
+         * edo.get.rotated([0,2,4,5,7],-1) //returns [7,0,2,4,5]
          * */
         rotated: (pitches,n) => {
             n=this.mod(n,pitches.length)
@@ -3162,28 +3166,6 @@ class EDO {
          * @example
          * //The quickest way to get to E (from 0) moving up with P5s and down with m3s
          * let path = edo.get.shortest_path(7,[5,-3]) // returns [ -3, 5, 5 ]
-         * let all = edo.get.permutations(path) //all permutation of the result
-         * //returns
-         * [
-         *  [ -3, 5, 5 ],
-         *  [ -3, 5, 5 ],
-         *  [ 5, -3, 5 ],
-         *  [ 5, 5, -3 ],
-         *  [ 5, -3, 5 ],
-         *  [ 5, 5, -3 ]
-         * ]
[…]
