### 2021-10-13 cbe5118
Added some Unit tests
