+++ test/unit.js
@@ -6,141 +6,238 @@ function unit_tests() {
     const edo17 = new EDO(17)
     const edo31 = new EDO(31)
     const edo1200 = new EDO(1200)
-    describe('EDO.convert', function () {
-        describe('Deterministic tests in 12EDO', function () {
-            let tests = {
-                cents_to_ratio: [edo12.convert.cents_to_ratio, [700], 1.4983070768766815],
-                cents_to_simple_ratio: [edo12.convert.cents_to_simple_ratio, [700], {
-                    "cents": 1901.9550008653873,
-                    "cents_in_octave": 701.9550008653873,
-                    "value": 3,
-                    "diff_in_octave": -1.9550008653873192,
-                    "ratio": "3:1",
-                    "original": 700
-                }],
-                interval_to_cents: [edo12.convert.interval_to_cents, [6], 600],
-                interval_to_ratio: [edo12.convert.interval_to_ratio, [7], 1.4983070768766815],
-                intervals_to_pitches: [edo12.convert.intervals_to_pitches, [[2, 3]], [0, 2, 5]],
-                intervals_to_scale: [edo12.convert.intervals_to_scale, [[2, 2, 1, 2, 2, 2, 1]], [0, 2, 4, 5, 7, 9, 11]],
-                midi_to_intervals: [edo12.convert.midi_to_intervals, [[60, 64, 57, 61]], [4, -7, 4]],
-                midi_to_name: [edo12.convert.midi_to_name, [[60, 62]], ["C4", "D4"]],
-                midi_to_freq: [edo12.convert.midi_to_freq, [[69, 70]], [440, 466.1637615180899]],
-                // name_to_scale: [edo.convert.name_to_scale,['12-1387']]
-                pc_to_name: [edo12.convert.pc_to_name, [4], "E"],
-                pitches_to_PCs: [edo12.convert.pitches_to_PCs, [[0, 2, 12, -2, 7]], [0, 2, 0, 10, 7]],
-                ratio_to_cents: [edo12.convert.ratio_to_cents, [5 / 4], 386.3137138648348],
-                ratio_to_interval: [edo12.convert.ratio_to_interval, [5 / 4, 20], [4]],
-                to_steps: [edo12.convert.to_steps, [[0, 2, 4, 5, 7, 9, 11]], [2, 2, 1, 2, 2, 2]]
 
-            }
+    const dia = edo12.scale([0,2,4,5,7,9,11]) //Diatonic scale in 12EDO
+    const WT = edo12.scale([0,2,4,6,8,10]) //Whole-tone scale
+    const penta =edo12.scale([0,2,4,7,9]) // The Pentatonic Scale
+    const octa = edo12.scale([0,2,3,5,6,8,9,11]) // The Octatonic scale
+    const maj3 = edo12.scale([0,4,7]) // The major triad
+    const min3 = edo12.scale([0,3,7]) // The minor triad
+    const f6f7 = edo12.scale([0,2,4,5,7,8,10]) // Major scale with flat 6 and flat 7
+    describe("EDO Class Unit Tests", function () {
+        describe('EDO.convert', function () {
+            describe('Deterministic tests in 12EDO', function () {
+                let tests = {
+                    cents_to_ratio: [edo12.convert.cents_to_ratio, [700], 1.4983070768766815],
[…]
