### 2021-10-13 bf395d9
changed PC to pitch-class in documentation
