+++ edo.js
@@ -664,7 +664,7 @@ class EDO {
             return PC[pc].trim()
         },
 
-        /** Normalizes any input to include PC only (to ignore octave displacement)
+        /** Normalizes any input to include pitch-classes only (to ignore octave displacement)
          * @param  {Array<Number>} pitches - any collection of pitches (e.g. a melody)
          * @returns {Array<Number>} the input as pitch classes
          * @memberOf EDO#convert
@@ -1008,7 +1008,7 @@ class EDO {
          * either 1 if note n is higher than n-1, 0 if note n is the same as n-1, and -1 if note n is lower than n-1
          * For instance <code>[0,0,4,7,4,7,4,0]</code> will in local mode will return <code>[0,1,1,-1,1,-1,-1]</code></p>
          *
-         * <p>If local is set to false (default), the contour of the line is expressed such that the actual pc class of the
+         * <p>If local is set to false (default), the contour of the line is expressed such that the actual pitch-class of the
          * note is removed but its relative position in regards to the entire line is kept.
          * <code>[0,4,7,12,16,7,12,16]</code> (Bach prelude in C) has 5 distinct note heights, so it will return
          * <code>[0,1,2,3, 4, 2,3, 4]</code> indicating the relative height of each note in the entire phrase</p>
@@ -1121,7 +1121,7 @@ class EDO {
             return this.edo - interval
         },
 
-        /** <p>Returns all the PCs of the EDO that the scale does not use.</p>
+        /** <p>Returns all the pitch-classes of the EDO that the scale does not use.</p>
          * @param {boolean} [from_0=false] - when true, the output will be normalized to 0.
          * @returns {Array<Number>}
          * @memberOf EDO#get
@@ -1847,7 +1847,7 @@ class EDO {
 
         /** Returns the inversion of a given set of pitches
          *
-         * @param  {Array<Number>} scale - a collection of pitches (not necessarily PCs)
+         * @param  {Array<Number>} scale - a collection of pitches (not necessarily pitch-classes)
          * @param  {Boolean} cache - if true, the result will be cached for faster retrival
          * @return {Array<Number>} The inverted input
          * @memberOf EDO#get
@@ -1973,7 +1973,7 @@ class EDO {
         //  *      E.g. If we consider C-E-G <code>(0,4,7)</code>, E and G appear as overtones of C at lower positions than C and G appear as overtones of E, and C and E as overtones of G.</p>
         //  *      <p>Note: a root can be highly dependent on context, therefore this algorithm at its current state cannot provide a decisive answer.</p>
         //  * @param  {Array<Number>} pitches - a collection of pitch classes
-        //  * @param  {Array<Number>} [limit=19] - The overtone limit by which PCs are approximated
+        //  * @param  {Array<Number>} [limit=19] - The overtone limit by which pitch-classes are approximated
         //  * @return {Number} The pitch-class of the likely root.
         //  * @memberOf EDO#get
         //  * @example
@@ -2087,7 +2087,7 @@ class EDO {
         },
 
         /** Returns the normal order of a given set of pitches
-         * @param  {Array<Number>} scale - a collection of PCs
+         * @param  {Array<Number>} scale - a collection of pitch-classes
          * @param  {Boolean} cache - if true, the result will be cached for faster retrieval
          * @return {Array<Array<Number>>}
          * @memberOf EDO#get
@@ -2302,7 +2302,7 @@ class EDO {
         },
 
         /** Returns the normal order of a given set of pitches
-         * @param  {Array<Number>} lst - a collection of PCs
+         * @param  {Array<Number>} lst - a collection of pitch-classes
          * @param  {Boolean} cache - if true, the result will be cached for faster retrival
          * @return {Array<Number>} The normal order of the input
          * @memberOf EDO#get
[…]
