+++ paper.bib
@@ -70,3 +70,23 @@
   volume={304},
   publisher={Yale University Press}
 }
+
+@inproceedings{music21,
+  author = {Cuthbert, Michael Scott and Ariza, Christopher},
+  year = 2010
+  title = {Music21: A Toolkit for Computer-Aided Musicology and Symbolic Music Data.},
+  added-at = {2020-03-12T00:00:00.000+0100},
+  biburl = {https://www.bibsonomy.org/bibtex/2655eb165e9efd7371a52cc1ce2d4efab/dblp},
+  booktitle = {ISMIR},
+  crossref = {conf/ismir/2010},
+  editor = {Downie, J. Stephen and Veltkamp, Remco C.},
+  ee = {http://ismir2010.ismir.net/proceedings/ismir2010-108.pdf},
+  interhash = {75ae23495cb8183efb69481e6ad51a5f},
+  intrahash = {655eb165e9efd7371a52cc1ce2d4efab},
+  isbn = {978-90-393-53813},
+  keywords = {dblp},
+  pages = {637-642},
+  publisher = {International Society for Music Information Retrieval},
+  timestamp = {2020-03-13T13:00:05.000+0100},
+  url = {http://dblp.uni-trier.de/db/conf/ismir/ismir2010.html#CuthbertA10},
+}
