+++ paper.bib
@@ -73,7 +73,7 @@
 
 @inproceedings{music21,
   author = {Cuthbert, Michael Scott and Ariza, Christopher},
-  year = 2010
+  year = {2010}
   title = {Music21: A Toolkit for Computer-Aided Musicology and Symbolic Music Data.},
   added-at = {2020-03-12T00:00:00.000+0100},
   biburl = {https://www.bibsonomy.org/bibtex/2655eb165e9efd7371a52cc1ce2d4efab/dblp},
