+++ paper.bib
@@ -1,6 +1,6 @@
 @book{Balzano:1982,
   Author = {Balzano, J. G.},
-  Year = 1982
+  Year = 1982,
   Title = {{The Pitch Set as a Level of Description for Studying Musical Pitch Perception}},
   url = {https://doi.org/10.1007/978-1-4684-8917-0_17},
   Booktitle = {Balzano G.J. (1982) The Pitch Set as a Level of Description for Studying Musical Pitch Perception. In: Clynes M. (eds) Music, Mind, and Brain. Springer, Boston, MA. https://doi.org/10.1007/978-1-4684-8917-0_17},
