### 2021-10-21 MichaelSel/edoJS#10 @napulen
[PAPER] Some minor corrections
This issue is related to the following review: openjournals/joss-reviews#3784

https://github.com/MichaelSel/edoJS/blob/feb06495b1ce4ed3a957ce2d4c1e08bb8fd67893/paper.md?plain=1#L32

- "western" -> "Western"; "non-western" -> "non-Western"
- "fundemental" -> "fundamental"

> EDO.js is by far the most comprehensive library that supports non-standard tuning to date

Any references/experiments to support the claim? Otherwise, it feels a bit hyperbolic. I'd tone it down a cent:

"EDO.js is, to the best of our knowledge, the most comprehensive library supporting non-standard tuning."
