### 2021-10-21 cabe1c9
changed PC to pitch-class in documentation
