### 2021-10-22 1f6a044
Added .ignore files and updated README.md with installation instructions (i.e., npm )
