### 2021-10-22 053f437
Updated paper to tone down statement
