### 2021-10-22 b0f1faf
Added build using rollup.config.js
