### 2021-10-22 20480a8
Corrected typos in paper.md
