### 2021-10-22 5a53703
Added better browser-server side compatibility Updated documentation
