+++ dist/edo.es.js
@@ -1,23 +1,3 @@
-/**
- * @author Michael Seltenreich <seltenmusic@gmail.com>
- * @version 1.2.0
- * @license
- * <pre>Copyright 2020 Michael Seltenreich
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.</pre>*/
-
-
-
 const environment = (typeof window === 'undefined') ? "server" : "browser";
 
 let fs, parseXML, midiParser;
@@ -359,6 +339,7 @@ class EDO {
     /**
      * Returns a new Scale Object with given pitches
      * @param  {Array<Number>} pitches - a collection of pitch classes
+     * @remarks "pitch classes" conform to the current tuning system used. 0-11 in 12EDO, 0-16 in 17EDO, etc.
      * @return {Scale}
      */
     scale(pitches) {
@@ -527,7 +508,7 @@ class EDO {
             return pitches
         },
 
-        /** Gets a list of intervals classes. Returns a scale as list of pitch classes.
+        /** Gets a series of intervallic units . Returns a scale as list of pitch classes
          *
          * @param  {Array<Number>} intervals - A list of intervals
          * @example
@@ -535,6 +516,7 @@ class EDO {
          * edo.convert.intervals_to_scale([2, 2, 1, 2, 2, 2, 1])
          * // returns [0,2,4,5,7,9,11]
          * @returns {Number} A scale made up by adding the intervals in order
+         * @remarks "pitch classes" according to the current tuning system used. 0-11 in 12EDO, 0-16 in 17EDO, etc.
          * @memberOf EDO#convert*/
         intervals_to_scale: (intervals) => {
             let pcs = [0];
@@ -667,6 +649,7 @@ class EDO {
          * @param  {Array<Number>} pitches - any collection of pitches (e.g. a melody)
          * @returns {Array<Number>} the input as pitch classes
          * @memberOf EDO#convert
+         * @remarks "pitch classes" according to the current tuning system used. 0-11 in 12EDO, 0-16 in 17EDO, etc.
          * @example
          * let edo = new EDO(12) // define a tuning system with 12 divisions of the octave
          * edo.convert.pitches_to_PCs([0,2,12,-2,7])
@@ -689,12 +672,13 @@ class EDO {
             return 1200 * Math.log2(ratio)
         },
 
-        /** Returns all of the IC in the EDO that equal to a given ratio (with a given tolerance in cents)
+        /** Returns all of the intervallic units in the EDO that equal to a given ratio (with a given tolerance in cents)
          *
          * @param  {Number} ratio - A harmonic ratio
          * @param  {Number} [tolerance=10] - a tolerance (allowed error) in cents
-         * @returns {Array<Number>} Interval classes that fit that ratio
+         * @returns {Array<Number>} Intervals that fit that ratio
          * @memberOf EDO#convert
+         * @remarks "intervallic units" are marked 0 t n-1 with n being equal to the current edo system. so 0-11 in 12EDO, and 0-16 in 17EDO
          * @example
          * let edo = new EDO(12) // define a tuning system
          * edo.convert.ratio_to_interval(3/2)
@@ -747,6 +731,7 @@ class EDO {
          * Returns the number of commons tones between two collections of pitches
          * @param  {Array<Number>} list1 - a collection of pitches (not necessarily pitch classes)
[…]

[…]
