+++ dist/edo.es.js
@@ -337,9 +337,9 @@ class EDO {
     }
 
     /**
-     * Returns a new Scale Object with given pitches
+     * <p>Returns a new Scale Object with given pitches</p>
+     * <p>Remark: "pitch classes" conform to the current tuning system used. 0-11 in 12EDO, 0-16 in 17EDO, etc.</p>
      * @param  {Array<Number>} pitches - a collection of pitch classes
-     * @remarks "pitch classes" conform to the current tuning system used. 0-11 in 12EDO, 0-16 in 17EDO, etc.
      * @return {Scale}
      */
     scale(pitches) {
@@ -508,15 +508,14 @@ class EDO {
             return pitches
         },
 
-        /** Gets a series of intervallic units . Returns a scale as list of pitch classes
-         *
+        /** <p>Gets a series of intervallic units . Returns a scale as list of pitch classes</p>
+         *<p>Remark: "pitch classes" conform to the current tuning system used. 0-11 in 12EDO, 0-16 in 17EDO, etc.</p>
          * @param  {Array<Number>} intervals - A list of intervals
          * @example
          * let edo = new EDO(12) // define a tuning system with 12 divisions of the octave
          * edo.convert.intervals_to_scale([2, 2, 1, 2, 2, 2, 1])
          * // returns [0,2,4,5,7,9,11]
          * @returns {Number} A scale made up by adding the intervals in order
-         * @remarks "pitch classes" according to the current tuning system used. 0-11 in 12EDO, 0-16 in 17EDO, etc.
          * @memberOf EDO#convert*/
         intervals_to_scale: (intervals) => {
             let pcs = [0];
@@ -645,11 +644,11 @@ class EDO {
             return PC[pc].trim()
         },
 
-        /** Normalizes any input to include pitch-classes only (to ignore octave displacement)
+        /** <p>Normalizes any input to include pitch-classes only (to ignore octave displacement)</p>
+         * <p>Remark: "pitch classes" conform to the current tuning system used. 0-11 in 12EDO, 0-16 in 17EDO, etc.</p>
          * @param  {Array<Number>} pitches - any collection of pitches (e.g. a melody)
          * @returns {Array<Number>} the input as pitch classes
          * @memberOf EDO#convert
-         * @remarks "pitch classes" according to the current tuning system used. 0-11 in 12EDO, 0-16 in 17EDO, etc.
          * @example
          * let edo = new EDO(12) // define a tuning system with 12 divisions of the octave
          * edo.convert.pitches_to_PCs([0,2,12,-2,7])
@@ -672,13 +671,12 @@ class EDO {
             return 1200 * Math.log2(ratio)
         },
 
-        /** Returns all of the intervallic units in the EDO that equal to a given ratio (with a given tolerance in cents)
-         *
+        /** <p>Returns all of the intervallic units in the EDO that equal to a given ratio (with a given tolerance in cents)</p>
+         *<p>Remark: "intervallic units" conform to the current tuning system used. E.g., 0-11 occupy 1 octave in 12EDO, 0-16 in 17EDO, etc.</p>
          * @param  {Number} ratio - A harmonic ratio
          * @param  {Number} [tolerance=10] - a tolerance (allowed error) in cents
          * @returns {Array<Number>} Intervals that fit that ratio
          * @memberOf EDO#convert
-         * @remarks "intervallic units" are marked 0 t n-1 with n being equal to the current edo system. so 0-11 in 12EDO, and 0-16 in 17EDO
          * @example
          * let edo = new EDO(12) // define a tuning system
          * edo.convert.ratio_to_interval(3/2)
@@ -729,9 +727,9 @@ class EDO {
     count = {
         /**
          * Returns the number of commons tones between two collections of pitches
[…]

[…]
