+++ index.js
@@ -1,0 +1,4 @@
+// index.js
+// entrpoint for NPM
+const edojs = require("./dist/edo.umd.js");
+module.exports = edojs;
