### 2021-11-01 #3784 @bmcfee
Hi @MichaelSel, I want to make sure we A) don't lose the thread on this, and B) are mindful of our reviewer's time and attention.  There's no rush or pressure on this from the JOSS side. If you have a sense of when you expect to have these issues resolved, we can have whedon set a reminder to come back in X weeks.  Thanks!
