### 2021-11-01 MichaelSel/edoJS#10 @napulen
On the last correction, a small typo `supports` -> `support`.
