### 2021-11-01 MichaelSel/edoJS#8 @napulen
The tests look good. I want to run them locally before closing the issue. Will try to do that tonight.
