### 2021-11-02 MichaelSel/edoJS#15 @napulen
[Code] Deprecated npm flag
This issue is related to the following review: openjournals/joss-reviews#3784

Running `npm install --dev` returns on

```
npm WARN install Usage of the `--dev` option is deprecated. Use `--include=dev` instead.
```
My impression is that `npm install --include=dev` works properly, so I'd verify this and replace the instructions in the `README`.

My environment is:
Ubuntu 20.04
node v16.13.0
npm v8.1.0
