### 2021-11-02 MichaelSel/edoJS#8 @napulen
Trying to run the tests locally, following the instructions of the `README`
```
npm install edo.js --dev
npm test
```

results in the following error:
```

> edo.js@1.2.16 test
> mocha

Error: Cannot find module './dist/edo.umd.js'
Require stack:
- /home/napulen/dev/edoJS/index.js
- /home/napulen/dev/edoJS/test/unit.js
    at Function.Module._resolveFilename (node:internal/modules/cjs/loader:933:15)
    at Function.Module._load (node:internal/modules/cjs/loader:778:27)
    at Module.require (node:internal/modules/cjs/loader:1005:19)
    at require (node:internal/modules/cjs/helpers:102:18)
    at Object.<anonymous> (/home/napulen/dev/edoJS/index.js:3:15)
    at Module._compile (node:internal/modules/cjs/loader:1101:14)
    at Object.Module._extensions..js (node:internal/modules/cjs/loader:1153:10)
    at Module.load (node:internal/modules/cjs/loader:981:32)
    at Function.Module._load (node:internal/modules/cjs/loader:822:12)
    at Module.require (node:internal/modules/cjs/loader:1005:19)
    at require (node:internal/modules/cjs/helpers:102:18)
    at unit_tests (/home/napulen/dev/edoJS/test/unit.js:2:17)
    at Object.<anonymous> (/home/napulen/dev/edoJS/test/unit.js:318:1)
    at Module._compile (node:internal/modules/cjs/loader:1101:14)
    at Object.Module._extensions..js (node:internal/modules/cjs/loader:1153:10)
    at Module.load (node:internal/modules/cjs/loader:981:32)
    at Function.Module._load (node:internal/modules/cjs/loader:822:12)
    at ModuleWrap.<anonymous> (node:internal/modules/esm/translators:190:29)
    at ModuleJob.run (node:internal/modules/esm/module_job:185:25)
    at async Promise.all (index 0)
    at async ESMLoader.import (node:internal/modules/esm/loader:281:24)
    at async importModuleDynamicallyWrapper (node:internal/vm/module:437:15)
    at async formattedImport (/home/napulen/dev/edoJS/node_modules/mocha/lib/nodejs/esm-utils.js:7:14)
    at async Object.exports.requireOrImport (/home/napulen/dev/edoJS/node_modules/mocha/lib/nodejs/esm-utils.js:48:32)
    at async Object.exports.loadFilesAsync (/home/napulen/dev/edoJS/node_modules/mocha/lib/nodejs/esm-utils.js:88:20)
    at async singleRun (/home/napulen/dev/edoJS/node_modules/mocha/lib/cli/run-helpers.js:125:3)
    at async Object.exports.handler (/home/napulen/dev/edoJS/node_modules/mocha/lib/cli/run.js:374:5)
```

This is on
- Ubuntu 20.04
- node v16.13.0
- npm v8.1.0
