### 2021-11-02 MichaelSel/edoJS#8 @napulen
There is no `dist/` folder generated by neither `npm install` nor `npm install --dev`.

After looking at the code, I noticed that the folder is created with the `rollup.config.js` script.

Using the documentation on [rollup](https://rollupjs.org/guide/en), I ran the config

```
npm install --global rollup
rollup --config rollup.config.js
```
This generated the `dist/` folder. After that, the tests pass

```
$ npm test

> edo.js@1.2.16 test
> mocha

  EDO Class Unit Tests
    EDO.convert
      Deterministic tests in 12EDO
        ✔ cents_to_ratio
        ✔ cents_to_simple_ratio
        ✔ interval_to_cents
        ✔ interval_to_ratio
        ✔ intervals_to_pitches
        ✔ intervals_to_scale
        ✔ midi_to_intervals
        ✔ midi_to_name
        ✔ midi_to_freq
        ✔ pc_to_name
        ✔ pitches_to_PCs
        ✔ ratio_to_cents
        ✔ ratio_to_interval
        ✔ to_steps
    EDO.count
      Deterministic tests in 12EDO
        ✔ common_tones
        ✔ differences
        ✔ pitches
    EDO.get
      Deterministic tests in 12EDO
        ✔ angle
        ✔ coordinates
        ✔ contour
        ✔ combinations
        ✔ complementary_interval
        ✔ complementary_set
        ✔ harp_pedals_to_pitches
        ✔ interval_class
        ✔ n_choose_k
        ✔ notes_from_cents
        ✔ sine_pair_dissonance
        ✔ resize_melody
        ✔ generated_scale
        ✔ generators
        ✔ intersection
        ✔ interval_traversed
        ✔ interval_stack
        ✔ inversion
        ✔ levenshtein
        ✔ maximal_rahn_difference
        ✔ maximal_carey_coherence_failures
        ✔ minimal_voice_leading
        ✔ modes
        ✔ necklace
        ✔ new_pitches
        ✔ ngrams
        ✔ normal_order
        ✔ stacked
        ✔ without
        ✔ partitioned_subsets
        ✔ permutations
        ✔ pitch_distribution
        ✔ pitch_fields
        ✔ ratio_approximation
        ✔ retrograde
        ✔ rotated
        ✔ rotations
        ✔ shortest_path
        ✔ starting_at
        ✔ subset_indices
        ✔ subsets
        ✔ transposition
        ✔ union
        ✔ unique_elements
        ✔ well_formed_scale
        ✔ without_chromatic_notes
    EDO.is
      Deterministic tests in 12EDO
        ✔ element_of
        ✔ rotation
        ✔ same
        ✔ subset
        ✔ transposition

  Scale Class Unit Tests
    Scale.count
      Deterministic tests in 12EDO
        ✔ chord_quality
        ✔ consecutive_steps
        ✔ imperfections
        ✔ count
        ✔ M3s
        ✔ m3s
        ✔ major_minor_triads
        ✔ modes
        ✔ n_chords
        ✔ P5s
        ✔ pitches
        ✔ rahn_differences
        ✔ rahn_contradictions
        ✔ rahn_ambiguities
        ✔ rotational_symmetries
        ✔ tetrachords
        ✔ thirds
        ✔ trichords
        ✔ unique_elements
    Scale.get
      Deterministic tests in 12EDO
        ✔ area
        ✔ diagnostic_intervals
        ✔ set_difference
        ✔ per_note_set_difference
        ✔ coordinates
        ✔ common_tone_transpositions
        ✔ complement
        ✔ chord_quality_from_shape
        ✔ melody_from_intervals
        ✔ generic_intervals
        ✔ specific_intervals
        ✔ interval_vector
        ✔ inversion
        ✔ lerdahl_attraction
        ✔ lerdahl_attraction_vector
        ✔ levenshtein
        ✔ myhill_property
        ✔ modes
        ✔ n_chords
        ✔ neighborhood
        ✔ normal_order
        ✔ permutations
        ✔ pitches
        ✔ position_of_quality
        ✔ prime_form
        ✔ product
        ✔ rotations
        ✔ rothenberg_propriety
        ✔ roughness
        ✔ sameness_quotient
        ✔ coherence_quotient
        ✔ scale_degree_transpositions
        ✔ segments
        ✔ sequence_transposition
        ✔ stacks
        ✔ step_sizes
        ✔ steps_to_qualities
        ✔ supersets
        ✔ tetrachords
        ✔ scale_degree_roles
        ✔ transposition
        ✔ transpositions_with_pitches
        ✔ trichords
        ✔ without
    Scale.is
      Deterministic tests in 12EDO
        ✔ deep
        ✔ distributionally_even
        ✔ in_lower_edos
        ✔ invertible
        ✔ maximally_even
        ✔ myhill_property
        ✔ mode_of
        ✔ MOLT
        ✔ normal_order
        ✔ one_of
        ✔ prime_form
        ✔ subset
    Scale.to
      Deterministic tests in 12EDO
        ✔ cents
        ✔ steps
```
