### 2021-11-02 MichaelSel/edoJS#8 @napulen
The `README` should include the instructions related to `rollup`, so that users can generate the distribution files to run the tests.
