### 2021-11-02 MichaelSel/edoJS#7 @napulen
The edits on the documentation look good. In my opinion, it reads better now.

This was a nice discussion! Closing now.
