### 2021-11-02 #3784 @napulen
I was unfortunately unable to run the tests locally without some further reading/scripts. It should be an easy fix though, I left the relevant comments in the issue https://github.com/MichaelSel/edoJS/issues/8, plus an additional issue about a deprecated `npm` flag.
