### 2021-11-02 MichaelSel/edoJS#8 @MichaelSel
You are correct, the README wasn't complete. I updated it now.

In fact you don't need to do the rollup, all you need to do is run
'''npm run build'''
to build all of the sources.

Then everything should run as expected.
