### 2021-11-02 #3784 @MichaelSel
I believe this is now resolved. I'll leave it up to Néstor to make the
final call.

On Tue, Nov 2, 2021 at 3:14 PM Néstor Nápoles López <
***@***.***> wrote:

> I was unfortunately unable to run the tests locally without some further
> reading/scripts. It should be an easy fix though, I left the relevant
> comments in the issue MichaelSel/edoJS#8
> <https://github.com/MichaelSel/edoJS/issues/8>, plus an additional issue
> about a deprecated npm flag.
>
