### 2021-11-02 MichaelSel/edoJS#8 @napulen
Confirming the build script runs on my end. Thanks!
