### 2021-11-02 #3784 @napulen
Confirming that all my issues have been resolved.
