### 2021-11-11 #3784 @albincorreya
@bmcfee Confirming all the issues except the automated CI pipeline has been resolved. Since the CI pipeline is not a must-have, I let @MichaelSel make the call. He will work on it during future iterations of the project. Thus not blocking this review as I understand.
