### 2021-11-12 #3784 @bmcfee
@MichaelSel things look generally good with the draft and the repo.  A couple of final to-dos for you:

1. Fix the missing DOI's in your references (see whedon comment above)
2. Create a new tag and release of the software incorporating changes made during the review.  (If it's already up to date, you can skip this step.)
3. Deposit the (new) release at an archive such as Zenodo or Figshare, and report back here with the DOI.  We'll use that to link the JOSS paper to the archived software version.
