+++ paper.bib
@@ -5,6 +5,7 @@
   url = {https://doi.org/10.1007/978-1-4684-8917-0_17},
   Booktitle = {Balzano G.J. (1982) The Pitch Set as a Level of Description for Studying Musical Pitch Perception. In: Clynes M. (eds) Music, Mind, and Brain. Springer, Boston, MA. https://doi.org/10.1007/978-1-4684-8917-0_17},
   Publisher = {Springer, Boston, MA},
+  doi = {10.1007/978-1-4684-8917-0_17},
 }
 
 @article{clampitt,
@@ -31,7 +32,8 @@
   volume={35},
   number={1/2},
   pages={33--60},
-  publisher={JSTOR}
+  publisher={JSTOR},
+  doi={10.2307/843809}
 }
 
 @article{vassilakis2001auditory,
@@ -42,14 +44,16 @@
   volume={110},
   number={5},
   pages={2755--2755},
-  publisher={Acoustical Society of America}
+  publisher={Acoustical Society of America},
+  doi = {10.1121/1.4777600}
 }
 
 @book{lerdahl2001tonal,
   author={Lerdahl, Fred},
   year={2001},
   title={Tonal pitch space},
-  publisher={Oxford University Press, USA}
+  publisher={Oxford University Press, USA},
+  doi = {10.1093/acprof:oso/9780195178296.001.0001}
 }
 
 @article{carey2007coherence,
@@ -60,7 +64,8 @@
   volume={1},
   number={2},
   pages={79--98},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi = {10.1080/17459730701376743}
 }
 
 @book{forte1973structure,
@@ -68,7 +73,8 @@
   year={1973},
   title={The structure of atonal music},
   volume={304},
-  publisher={Yale University Press}
+  publisher={Yale University Press},
+  
 }
 
 @inproceedings{music21,
