+++ paper.bib
@@ -74,6 +74,8 @@
   title={The structure of atonal music},
   volume={304},
   publisher={Yale University Press},
+  isbn = {9780300021202},
+  
   
 }
 
