### 2021-11-16 #3784 @MichaelSel
@bmcfee, Hi Brian.
I updated the DOIs of everything with the exception of Allan Forte's book "The Structure of Tonal Music" which doesn't seem to have a DOI listen anywhere. I did include its ISBN instead. The DOI whedon proposes is the wrong one.

I put a release on the repository and uploaded that to Zenodo. The DOI generated for that is 10.5281/zenodo.5705161
Here is the Zenodo record: https://zenodo.org/record/5705161

Please let me know if there anything you like me to do.

Thanks!
