### 2021-11-16 #3784 @bmcfee
@MichaelSel thanks!

It looks like you're missing a DOI for music21, which ought to be https://doi.org/10.5281/zenodo.1416113

Everything else looks okay to me.
