+++ paper.bib
@@ -97,4 +97,5 @@
   publisher = {International Society for Music Information Retrieval},
   timestamp = {2020-03-13T13:00:05.000+0100},
   url = {http://dblp.uni-trier.de/db/conf/ismir/ismir2010.html#CuthbertA10},
+  doi = {10.5281/zenodo.1416114}
 }
